refactor(game-engine): extract move sub-handlers (S27.8.6)
Slice 6 de S27.8 (refactor monolith `actions.ts`).

Extraction (`actions/move-handlers.ts`)
- `handleDodgeRoll` : esquive complete (modifiers + skills, relance
  Dodge / Break Tackle, GFI surajoute si PM=0, Shadowing
  post-esquive, touchdown, ramassage, Arm Bar sur l'echec).
- `handleNormalMove` : deplacement standard (decremente PM, ou GFI
  si PM=0 avec relance Sure Feet, touchdown, pickup).

Aucune dependance interne au dispatcher : tout passe par les
modules deja extraits (`failure-helpers`, `ball-pickup`) +
`core/game-state` (canUseTeamReroll, hasPlayerActed, ...) +
`mechanics/*` + `skills/*`.

`handleMove` reste dans `actions.ts` car il consomme
`getLegalMoves` (encore couple a `actions.ts`). Il importe
`handleDodgeRoll` et `handleNormalMove` depuis le nouveau module.

Resultat
- `actions.ts` 2135 -> 1878 lignes (-257 nettes ; cumul -828 vs
  origine 2706 = ~30.6%).
- `move-handlers.ts` ajoute a la liste d'exclusion coverage.
- 4888 tests game-engine passent (incluant le test integration
  Arm Bar via dodge failure que j'avais initialement casse en
  oubliant de propager `getArmBarBonus(next, player, from)` au
  call-site `applyRollFailure` ; corrige).
- Coverage 25.33% maintenue.

Reste pour atteindre la cible <= 600 lignes :
- handleLeap (depend de getLegalMoves, cycle a casser)
- handleDodge / handleMove (couplage getLegalMoves)
- handleBlock / choice handlers
- handleBlitz
