+++ packages/game-engine/src/actions/actions.ts
@@ -126,6 +126,13 @@ import {
 // pour servir de feuille reutilisable par handleNormalMove /
 // handleDodgeRoll / handleLeap / handleRerollChoose.
 import { handleBallPickup } from './ball-pickup';
+// S27.8.6 — `handleDodgeRoll` et `handleNormalMove` extraits dans
+// `actions/move-handlers.ts`. Appeles depuis `handleMove` (qui reste
+// dans ce fichier car couple a `getLegalMoves`).
+import {
+  handleDodgeRoll,
+  handleNormalMove,
+} from './move-handlers';
 import { canDumpOff, getDumpOffReceivers, executeDumpOff } from '../mechanics/dump-off';
 import { checkDauntless } from '../mechanics/dauntless';
 import { checkBreakTackle } from '../mechanics/break-tackle';
@@ -914,274 +921,10 @@ function handleMove(
   }
 }
 
-/**
- * Gère un jet d'esquive
- */
-function handleDodgeRoll(
-  state: GameState,
-  player: Player,
-  from: Position,
-  to: Position,
-  rng: RNG,
-  idx: number
-): GameState {
-  // Calculer les modificateurs de désquive (malus pour adversaires à l'arrivée + skills)
-  const baseDodgeModifiers = calculateDodgeModifiers(state, from, to, player.team);
-  const skillDodgeModifiers = getDodgeSkillModifiers(state, player, from);
-  const dodgeModifiers = baseDodgeModifiers + skillDodgeModifiers;
-
-  // Effectuer le jet d'esquive avec les modificateurs
-  const dodgeResult = performDodgeRollWithNotification(player, rng, dodgeModifiers);
-
-  let next = structuredClone(state) as GameState;
-  next.lastDiceResult = dodgeResult;
-
-  // Log du jet d'esquive
-  const logEntry = createLogEntry(
-    'dice',
-    `Jet d'esquive: ${dodgeResult.diceRoll}/${dodgeResult.targetNumber} ${dodgeResult.success ? '✓' : '✗'}`,
-    player.id,
-    player.team,
-    {
-      diceRoll: dodgeResult.diceRoll,
-      targetNumber: dodgeResult.targetNumber,
-      success: dodgeResult.success,
-      modifiers: dodgeModifiers,
-    }
-  );
-  next.gameLog = [...next.gameLog, logEntry];
-
-  // Le joueur se déplace toujours, que le jet d'esquive réussisse ou échoue
-  const isDodgeGFI = next.players[idx].pm <= 0;
-  next.players[idx].pos = { ...to };
-  if (isDodgeGFI) {
-    next.players[idx].gfiUsed = (next.players[idx].gfiUsed ?? 0) + 1;
-  } else {
-    next.players[idx].pm = Math.max(0, next.players[idx].pm - 1);
-  }
-
-  // Enregistrer l'action de mouvement seulement si c'est le premier mouvement
-  if (!hasPlayerActed(next, player.id)) {
-    next = setPlayerAction(next, player.id, 'MOVE');
-  }
-
-  // Vérifier si le tour du joueur doit se terminer
-  next = checkPlayerTurnEnd(next, player.id);
-
-  // Shadowing : les adversaires avec Shadowing adjacents à la case quittée
-  // tentent de suivre le dodger. Résolu une seule fois par esquive, après
-  // que le joueur ait bougé et indépendamment du résultat final (BB3).
-  next = resolveShadowingAfterDodge(next, player, from, rng);
-
-  // Break Tackle (BB3): une fois par activation, après un Dodge raté, ajoute
-  // +1 (ST <= 4) ou +2 (ST >= 5) au jet. Appliqué avant toute relance.
-  let finalDodgeSuccess = dodgeResult.success;
-  if (!finalDodgeSuccess) {
-    const breakTackleCheck = checkBreakTackle(
-      next,
-      next.players[idx],
-      dodgeResult.diceRoll,
-      dodgeResult.targetNumber,
-      dodgeResult.success
-    );
-    if (breakTackleCheck.triggered) {
-      next = breakTackleCheck.newState;
-      finalDodgeSuccess = true;
-    }
-  }
-  if (!finalDodgeSuccess && canSkillReroll(player, 'on-dodge', state)) {
-    const rerollLog = createLogEntry('dice', `Dodge : relance de l'esquive (${dodgeResult.diceRoll} raté)`, player.id, player.team);
-    next.gameLog = [...next.gameLog, rerollLog];
-    const rerollResult = performDodgeRollWithNotification(player, rng, dodgeModifiers);
-    next.lastDiceResult = rerollResult;
-    const rerollLogEntry = createLogEntry(
-      'dice',
-      `Relance esquive: ${rerollResult.diceRoll}/${rerollResult.targetNumber} ${rerollResult.success ? '✓' : '✗'}`,
-      player.id, player.team,
-      { diceRoll: rerollResult.diceRoll, targetNumber: rerollResult.targetNumber, success: rerollResult.success }
-    );
-    next.gameLog = [...next.gameLog, rerollLogEntry];
-    finalDodgeSuccess = rerollResult.success;
-  }
-
-  if (finalDodgeSuccess) {
-    // Si c'est aussi un GFI, jet supplémentaire de GFI (2+ sur D6)
-    if (isDodgeGFI) {
-      let gfiRoll = Math.floor(rng() * 6) + 1;
-      let gfiSuccess = gfiRoll >= 2;
-
-      // Sure Feet auto-reroll (via skill registry)
-      if (!gfiSuccess && canSkillReroll(player, 'on-gfi', state)) {
-        const sfLog = createLogEntry('dice', `Sure Feet : relance du GFI (${gfiRoll} raté)`, player.id, player.team);
-        next.gameLog = [...next.gameLog, sfLog];
-        gfiRoll = Math.floor(rng() * 6) + 1;
-        gfiSuccess = gfiRoll >= 2;
-      }
-
-      const gfiLogEntry = createLogEntry(
-        'dice',
-        `GFI (Going For It) après esquive: ${gfiRoll}/2 ${gfiSuccess ? '✓' : '✗'}`,
-        player.id, player.team,
-        { diceRoll: gfiRoll, targetNumber: 2, success: gfiSuccess }
-      );
-      next.gameLog = [...next.gameLog, gfiLogEntry];
-      next.lastDiceResult = { type: 'dodge' as any, playerId: player.id, diceRoll: gfiRoll, targetNumber: 2, success: gfiSuccess, modifiers: 0, playerName: player.name };
-
-      if (!gfiSuccess) {
-        // Offrir relance d'équipe si disponible
-        if (canUseTeamReroll(next, player.team)) {
-          next.pendingReroll = { rollType: 'gfi', playerId: player.id, team: player.team, targetNumber: 2, modifiers: 0, playerIndex: idx, to };
-          return next;
-        }
-        return applyRollFailure(next, idx, rng);
-      }
-    }
-
-    // Si le joueur porte la balle et atteint l'en-but adverse -> touchdown
-    const mover = next.players[idx];
-    if (mover.hasBall && isInOpponentEndzone(next, mover)) {
-      return awardTouchdown(next, mover.team, mover);
-    }
-  } else {
-    // Esquive échouée (après skill reroll éventuel) : offrir relance d'équipe
-    if (canUseTeamReroll(next, player.team)) {
-      next.pendingReroll = { rollType: 'dodge', playerId: player.id, team: player.team, targetNumber: dodgeResult.targetNumber, modifiers: dodgeModifiers, playerIndex: idx, from, to };
-      return next;
-    }
-    // Pas de relance disponible : appliquer l'echec, avec bonus Arm Bar
-    // si un adversaire adjacent a la case d'origine possede ce skill.
-    return applyRollFailure(next, idx, rng, getArmBarBonus(next, player, from));
-  }
-
-  return next;
-}
-
-/**
- * Gère un mouvement normal (sans jet d'esquive)
- */
-function handleNormalMove(
-  state: GameState,
-  player: Player,
-  from: Position,
-  to: Position,
-  rng: RNG,
-  idx: number
-): GameState {
-  let next = structuredClone(state) as GameState;
-  const isGFI = next.players[idx].pm <= 0;
-
-  // Déplacer le joueur
-  next.players[idx].pos = { ...to };
-
-  if (isGFI) {
-    // GFI : ne décrémente pas pm, incrémente gfiUsed
-    next.players[idx].gfiUsed = (next.players[idx].gfiUsed ?? 0) + 1;
-
-    // Jet de GFI : 2+ sur D6
-    let gfiRoll = Math.floor(rng() * 6) + 1;
-    let gfiSuccess = gfiRoll >= 2;
-
-    // Sure Feet : relance automatique du GFI raté (via skill registry)
-    if (!gfiSuccess && canSkillReroll(player, 'on-gfi', state)) {
-      const rerollLog = createLogEntry(
-        'dice',
-        `Sure Feet : relance du GFI (${gfiRoll} raté)`,
-        player.id,
-        player.team
-      );
-      next.gameLog = [...next.gameLog, rerollLog];
-      gfiRoll = Math.floor(rng() * 6) + 1;
-      gfiSuccess = gfiRoll >= 2;
-    }
-
-    const gfiLogEntry = createLogEntry(
-      'dice',
-      `GFI (Going For It): ${gfiRoll}/2 ${gfiSuccess ? '✓' : '✗'}`,
-      player.id,
-      player.team,
-      { diceRoll: gfiRoll, targetNumber: 2, success: gfiSuccess }
-    );
-    next.gameLog = [...next.gameLog, gfiLogEntry];
-    next.lastDiceResult = {
-      type: 'dodge' as any,
-      playerId: player.id,
-      diceRoll: gfiRoll,
-      targetNumber: 2,
-      success: gfiSuccess,
-      modifiers: 0,
-      playerName: player.name,
-    };
-
-    // Enregistrer l'action
-    if (!hasPlayerActed(next, player.id)) {
-      next = setPlayerAction(next, player.id, 'MOVE');
-    }
-
-    if (!gfiSuccess) {
-      // GFI échoué : offrir relance d'équipe si disponible
-      if (canUseTeamReroll(next, player.team)) {
-        next.pendingReroll = { rollType: 'gfi', playerId: player.id, team: player.team, targetNumber: 2, modifiers: 0, playerIndex: idx, to };
-        return next;
-      }
-      // Pas de relance : appliquer l'échec
-      return applyRollFailure(next, idx, rng);
-    }
-
-    // GFI réussi : continuer normalement
-    next = checkPlayerTurnEnd(next, player.id);
-
-    // Touchdown check
-    const mover = next.players[idx];
-    if (mover.hasBall && isInOpponentEndzone(next, mover)) {
-      return awardTouchdown(next, mover.team, mover);
-    }
-
-    // Ramassage de balle
-    if (next.ball && samePos(next.ball, to)) {
-      return handleBallPickup(next, player, rng, idx);
-    }
-
-    return next;
-  }
-
-  // Mouvement normal (a des PM)
-  next.players[idx].pm = Math.max(0, next.players[idx].pm - 1);
-
-  // Enregistrer l'action de mouvement seulement si c'est le premier mouvement
-  if (!hasPlayerActed(next, player.id)) {
-    next = setPlayerAction(next, player.id, 'MOVE');
-  }
-
-  // Vérifier si le tour du joueur doit se terminer
-  next = checkPlayerTurnEnd(next, player.id);
-
-  // Log du mouvement
-  const moveLogEntry = createLogEntry(
-    'action',
-    `Mouvement vers (${to.x}, ${to.y})`,
-    player.id,
-    player.team
-  );
-  next.gameLog = [...next.gameLog, moveLogEntry];
-  // Réinitialiser le résultat de dés après un mouvement normal
-  next.lastDiceResult = undefined;
-
-  // Si le joueur porte la balle et atteint l'en-but adverse -> touchdown
-  const mover = next.players[idx];
-  if (mover.hasBall && isInOpponentEndzone(next, mover)) {
-    return awardTouchdown(next, mover.team, mover);
-  }
-
-  // Ramassage de balle avec jet d'agilité
-  if (next.ball && samePos(next.ball, to)) {
-    return handleBallPickup(next, player, rng, idx);
-  }
-
-  return next;
-}
-
 // S27.8.5 — `handleBallPickup` extrait dans `actions/ball-pickup.ts`
 // (re-importe en haut de ce fichier).
+// S27.8.6 — `handleDodgeRoll` et `handleNormalMove` extraits dans
+// `actions/move-handlers.ts` (re-importes en haut de ce fichier).
 
 /**
  * Gère une action d'esquive explicite
