+++ packages/game-engine/vitest.config.ts
@@ -38,6 +38,7 @@ export default defineConfig({
         'src/actions/turn-foul-actions.ts',
         'src/actions/failure-helpers.ts',
         'src/actions/ball-pickup.ts',
+        'src/actions/move-handlers.ts',
       ],
       // S25.4 — Seuils initiaux conservateurs. v8 reporte 0% lines sur
       // certains fichiers TS du moteur a cause d'un quirk source-map ;
